integration: move 'model expect utils' functions
